[ut] added further description of the workspace bucket, have lint ignore trailing whitespace
untested as upload script appears to be hanging, but passed lint if W291 is added. W291 is added because trailing double whitespace is required for a line break in markdown, which is needed towards the beginning in the assumptions section. The trailing whitespace "issue" does not in any way effect Python parsing as indentation issues are parsed seperately, so it was never really needed anyway besides to keep things tidy. As for the workspace bucket changes, subdirectory is now a variable to allow the user to see the relevent part of their workspace bucket and create the cram-crai table without having to change two seperate points in the notebook. There is also a brief note on empty subdirectories, as this trips up a lot of first time users. The fact subdirectories do not actually exist is obscured to avoid confusion.